Reference

DiffOpt.Dπ

DiffOpt.Dπ

DiffOpt.Dπ

DiffOpt.Dπ

DiffOpt.Dπ

DiffOpt.backward!

DiffOpt.backward_conic!

DiffOpt.create_LHS_matrix

DiffOpt.create_RHS_matrix

DiffOpt.generate_lp

DiffOpt.generate_qp

DiffOpt.get_problem_data

DiffOpt.unvec_symm

DiffOpt.vec_symm

DiffOpt.π

DiffOpt.π

DiffOpt.π

DiffOpt.π

DiffOpt.π

DiffOpt.Dπ — Method

derivative of projection of vector `z` on a product of cones

DiffOpt.Dπ — Method

derivative of projection of vector `z` on Nonnegative cone i.e. K = R+

DiffOpt.Dπ — Method

derivative of projection of vector `z` on positive semidefinite cone i.e. K = S^n⨥

DiffOpt.Dπ — Method

derivative of projection of vector `z` on second order cone i.e. K = {(t, x) ∈ R+ × Rn |  ||x|| ≤ t }

DiffOpt.Dπ — Method

derivative of projection of vector `z` on zero cone i.e. K = {0}

DiffOpt.backward! — Method

Method to differentiate optimal solution `z` and return
product of jacobian matrices (`dz / dQ`, `dz / dq`, etc) with 
the backward pass vector `dl / dz`

The method computes the product of 
1. jacobian of problem solution `z*` with respect to 
    problem parameters `params` recieved as method arguments
2. a backward pass vector `dl / dz`, where `l` can be a loss function

Note that does not returns the actual jacobians

For more info refer eqn(7) and eqn(8) of https://arxiv.org/pdf/1703.00443.pdf

DiffOpt.backward_conic! — Method

Method to differentiate optimal solution `x`, `y`, `s`

DiffOpt.create_LHS_matrix — Function

Inverse matrix specified on RHS of eqn(7) in https://arxiv.org/pdf/1703.00443.pdf

DiffOpt.create_RHS_matrix — Function

Right hand side of eqn(6) in https://arxiv.org/pdf/1703.00443.pdf

DiffOpt.generate_lp — Method

Generates a non-trivial random MOI linear program by adding variables and constraints to MOI compatible Optimizer optimizer

minimize c' * x subject to Ax <= b, x >= 0 where x in R^{n}, A in R^{m*n}, b in R^{m}, c in R^{n}

Note: Mutates the optimizer object

DiffOpt.generate_qp — Method

Generates a non-trivial random MOI convex quadratic program by adding variables and constraints to MOI compatible Optimizer optimizer

minimize 0.5 * x' * Q * x + q' * x subject to Gx <= h, Ax == b where x in R^{n}, Q in R^{n*n}, q in R^{n}, G in R^{m,n}, h in R^{m}, A in R^{p*n}, b in R^{p}

Note: (1) Mutates the optimizer object (2) Matrix Q is Positive Semidefinite

DiffOpt.get_problem_data — Method

Return problem parameters as matrices along with other problem info

DiffOpt.unvec_symm — Method

Returns a dim-by-dim symmetric matrix corresponding to `x`.

`x` is a vector of length dim*(dim + 1)/2, corresponding to a symmetric
matrix; the correspondence is as in SCS.
X = [ X11 X12 ... X1k
      X21 X22 ... X2k
      ...
      Xk1 Xk2 ... Xkk ],
where
vec(X) = (X11, sqrt(2)*X21, ..., sqrt(2)*Xk1, X22, sqrt(2)*X32, ..., Xkk)

DiffOpt.vec_symm — Method

Returns a vectorized representation of a symmetric matrix `X`.
Vectorization (including scaling) as per SCS.
vec(X) = (X11, sqrt(2)*X21, ..., sqrt(2)*Xk1, X22, sqrt(2)*X32, ..., Xkk)

DiffOpt.π — Function

projection of vector `z` on zero cone i.e. K = {0} or its dual

DiffOpt.π — Method

Projection onto R^n x K^* x R_+
`cones` represents a convex cone K, and K^* is its dual cone

DiffOpt.π — Method

projection of vector `z` on Nonnegative cone i.e. K = R+

DiffOpt.π — Method

projection of vector `z` on positive semidefinite cone i.e. K = S^n⨥

DiffOpt.π — Method

projection of vector `z` on second order cone i.e. K = {(t, x) ∈ R+ × Rn |  ||x|| ≤ t }