Reference

DiffOpt.BackwardInVariablePrimal

DiffOpt.BackwardOut

DiffOpt.BackwardOutObjective

DiffOpt.ConstraintCoefficient

DiffOpt.ConstraintConstant

DiffOpt.ForwardInConstraint

DiffOpt.ForwardInObjective

DiffOpt.ForwardOutVariablePrimal

DiffOpt.Dπ

DiffOpt._backward_conic

DiffOpt._backward_quad

DiffOpt._forward_conic

DiffOpt._forward_quad

DiffOpt.backward

DiffOpt.create_LHS_matrix

DiffOpt.diff_optimizer

DiffOpt.diff_optimizer

DiffOpt.forward

DiffOpt.get_problem_data

DiffOpt.map_rows

DiffOpt.π

DiffOpt.BackwardInVariablePrimal — Type

BackwardInVariablePrimal

A MOI.AbstractVariableAttribute to set input data to backward differentiation, that is, problem solution. T

For instance, to set the tangent of the variable of index vi, do the following:

MOI.set(model, DiffOpt.BackwardInVariablePrimal(), x)

DiffOpt.BackwardOut — Type

BackwardOut{T}

A AbstractDiffAttribute to get output data to backward differentiation, that is, problem solution. The solution data includes: ConstraintConstant and ConstraintCoefficient. The latter can only be used in linearly constrained quadratic models.

MOI.get(model, DiffOpt.BackwardOut{DiffOpt.ConstraintCoefficient}(), x)

DiffOpt.BackwardOutObjective — Type

BackwardOutObjective

A MOI.AbstractModelAttribute to get output data to backward differentiation, that is, problem input data.

MOI.get(model, DiffOpt.BackwardOutObjective)

DiffOpt.ConstraintCoefficient — Type

ConstraintCoefficient

An attribute to set input and get output differentials from forward and backward differentiation related to the linear coefficient associated to a pair: MOI.VariableIndex and MOI.ConstraintIndex.

DiffOpt.ConstraintConstant — Type

ConstraintConstant

An attribute to set input and get output differentials from forward and backward differentiation related to the constant term associated to a MOI.ConstraintIndex.

DiffOpt.ForwardInConstraint — Type

ForwardInConstraint

A MOI.AbstractConstraintAttribute to set input data to forward differentiation, that is, problem input data.

For instance, if the scalar constraint of index ci contains θ * (x + 2y), for the purpose of computinig the derivative with respect to θ, the following should be set:

fx = MOI.SingleVariable(x)
fy = MOI.SingleVariable(y)
MOI.set(model, DiffOpt.ForwardInConstraint(), ci, 1.0 * fx + 2.0 * fy)

DiffOpt.ForwardInObjective — Type

ForwardInObjective

A MOI.AbstractModelAttribute to set input data to forward differentiation, that is, problem input data. The possible values are any MOI.AbstractScalarFunction. A MOI.ScalarQuadraticFunction can only be used in linearly constrained quadratic models.

For instance, if the objective contains θ * (x + 2y), for the purpose of computinig the derivative with respect to θ, the following should be set:

fx = MOI.SingleVariable(x)
fy = MOI.SingleVariable(y)
MOI.set(model, DiffOpt.ForwardInObjective(), 1.0 * fx + 2.0 * fy)

where x and y are the relevant MOI.VariableIndex.

DiffOpt.ForwardOutVariablePrimal — Type

ForwardOutVariablePrimal

A MOI.AbstractVariableAttribute to get output data from forward differentiation, that is, problem solution.

For instance, to get the tangent of the variable of index vi corresponding to the tangents given to ForwardInObjective and ForwardInConstraint, do the following:

MOI.get(model, DiffOpt.ForwardOutVariablePrimal(), vi)

DiffOpt.Dπ — Method

Dπ(v::Vector{Float64}, model, conic_form::MatOI.GeometricConicForm, index_map::MOIU.IndexMap)

Given a model, its conic_form and the index_map from the indices of model to the indices of conic_form, find the gradient of the projection of the vectors v of length equal to the number of rows in the conic form onto the cartesian product of the cones corresponding to these rows. For more info, refer to https://github.com/matbesancon/MathOptSetDistances.jl

DiffOpt._backward_conic — Method

_backward_conic(model::Optimizer, dx::Vector{Float64}, dy::Vector{Float64}, ds::Vector{Float64})

Method to compute the product of the transpose of the derivative (Jacobian) at the conic program parameters A, b, c to the perturbations dx, dy, ds. This is similar to backward.

For theoretical background, refer Section 3 of Differentiating Through a Cone Program, https://arxiv.org/abs/1904.09043

DiffOpt._backward_quad — Method

_backward_quad(model::Optimizer)

Method to differentiate optimal solution z and return product of jacobian matrices (dz / dQ, dz / dq, etc) with the backward pass vector dl / dz

The method computes the product of

jacobian of problem solution z* with respect to problem parameters set with the BackwardInVariablePrimal

a backward pass vector dl / dz, where l can be a loss function

Note that this method does not returns the actual jacobians.

For more info refer eqn(7) and eqn(8) of https://arxiv.org/pdf/1703.00443.pdf

DiffOpt._forward_conic — Method

_forward_conic(model::Optimizer)

Method to compute the product of the derivative (Jacobian) at the conic program parameters A, b, c to the perturbations dA, db, dc. This is similar to forward.

For theoretical background, refer Section 3 of Differentiating Through a Cone Program, https://arxiv.org/abs/1904.09043

DiffOpt._forward_quad — Method

_forward_quad(model::Optimizer)

DiffOpt.backward — Method

backward(model::Optimizer)

Wrapper method for the backward pass. This method will consider as input a currently solved problem and differentials with respect to the solution set with the BackwardInVariablePrimal attribute. The output problem data differentials can be queried with the attribute BackwardOut.

DiffOpt.create_LHS_matrix — Function

create_LHS_matrix(z, λ, Q, G, h, A=nothing)

Inverse matrix specified on RHS of eqn(7) in https://arxiv.org/pdf/1703.00443.pdf

Helper method while calling _backward_quad

DiffOpt.diff_optimizer — Method

diff_optimizer(optimizer_constructor)::Optimizer

Creates a DiffOpt.Optimizer, which is an MOI layer with an internal optimizer and other utility methods. Results (primal, dual and slack values) are obtained by querying the internal optimizer instantiated using the optimizer_constructor. These values are required for find jacobians with respect to problem data.

One define a differentiable model by using any solver of choice. Example:

julia> using DiffOpt, GLPK

julia> model = diff_optimizer(GLPK.Optimizer)
julia> model.add_variable(x)
julia> model.add_constraint(...)

julia> _backward_quad(model)  # for convex quadratic models

julia> _backward_quad(model)  # for convex conic models

DiffOpt.forward — Method

forward(model::Optimizer)

Wrapper method for the forward pass. This method will consider as input a currently solved problem and differentials with respect to problem data set with the ForwardInObjective and ForwardInConstraint attributes. The output solution differentials can be queried with the attribute ForwardOutVariablePrimal.

DiffOpt.get_problem_data — Method

get_problem_data(model::MOI.AbstractOptimizer)

Return problem parameters as matrices along with other program info such as number of constraints, variables, etc

DiffOpt.map_rows — Method

map_rows(f::Function, model, conic_form::MatOI.GeometricConicForm, index_map::MOIU.IndexMap, map_mode::Union{Nested{T}, Flattened{T}})

Given a model, its conic_form, the index_map from the indices of model to the indices of conic_form and map_mode of type Nested (resp. Flattened), return a Vector{T} of length equal to the number of cones (resp. rows) in the conic form where the value for the index (resp. rows) corresponding to each cone is equal to f(ci, r) where ci is the corresponding constraint index in model and r is a UnitRange of the corresponding rows in the conic form.

DiffOpt.π — Method

π(v::Vector{Float64}, model::MOI.ModelLike, conic_form::MatOI.GeometricConicForm, index_map::MOIU.IndexMap)

Given a model, its conic_form and the index_map from the indices of model to the indices of conic_form, find the projection of the vectors v of length equal to the number of rows in the conic form onto the cartesian product of the cones corresponding to these rows. For more info, refer to https://github.com/matbesancon/MathOptSetDistances.jl