Reference

DiffOpt.Dπ

DiffOpt.backward!

DiffOpt.backward_conic!

DiffOpt.create_LHS_matrix

DiffOpt.generate_lp

DiffOpt.generate_qp

DiffOpt.get_problem_data

DiffOpt.π

DiffOpt.Dπ — Method

Dπ(cones::Array{<: MOI.AbstractVectorSet}, v::Vector{Vector{Float64}})

Find gradient of projection of vectors in v on product of cones. For more info, refer https://github.com/matbesancon/MathOptSetDistances.jl

DiffOpt.backward! — Method

backward!(model::Optimizer, params::Array{String}, dl_dz::Array{Float64})

Method to differentiate optimal solution z and return product of jacobian matrices (dz / dQ, dz / dq, etc) with the backward pass vector dl / dz

The method computes the product of

jacobian of problem solution z* with respect to problem parameters params recieved as method arguments

a backward pass vector dl / dz, where l can be a loss function

Note that this method does not returns the actual jacobians.

For more info refer eqn(7) and eqn(8) of https://arxiv.org/pdf/1703.00443.pdf

DiffOpt.backward_conic! — Method

backward_conic!(model::Optimizer, dA::Array{Float64,2}, db::Array{Float64}, dc::Array{Float64})

Method to differentiate optimal solution x, y, s given perturbations related to conic program parameters A, b, c. This is similar to backward! method but it this does returns the actual jacobians.

For theoretical background, refer Section 3 of Differentiating Through a Cone Program, https://arxiv.org/abs/1904.09043

DiffOpt.create_LHS_matrix — Function

create_LHS_matrix(z, λ, Q, G, h, A=nothing)

Inverse matrix specified on RHS of eqn(7) in https://arxiv.org/pdf/1703.00443.pdf

Helper method while calling backward!

DiffOpt.generate_lp — Method

Generates a non-trivial random MOI linear program by adding variables and constraints to MOI compatible Optimizer optimizer

minimize c' * x subject to Ax <= b, x >= 0 where x in R^{n}, A in R^{m*n}, b in R^{m}, c in R^{n}

Note: Mutates the optimizer object

DiffOpt.generate_qp — Method

Generates a non-trivial random MOI convex quadratic program by adding variables and constraints to MOI compatible Optimizer optimizer

minimize 0.5 * x' * Q * x + q' * x subject to Gx <= h, Ax == b where x in R^{n}, Q in R^{n*n}, q in R^{n}, G in R^{m,n}, h in R^{m}, A in R^{p*n}, b in R^{p}

Note: (1) Mutates the optimizer object (2) Matrix Q is Positive Semidefinite

DiffOpt.get_problem_data — Method

get_problem_data(model::MOI.AbstractOptimizer)

Return problem parameters as matrices along with other program info such as number of constraints, variables, etc

DiffOpt.π — Method

π(cones::Array{<: MOI.AbstractVectorSet}, v::Vector{Vector{Float64}})

Find projection of vectors in v on product of cones. For more info, refer https://github.com/matbesancon/MathOptSetDistances.jl