Reference

DiffOpt.AbstractLazyScalarFunction

DiffOpt.BackwardInVariablePrimal

DiffOpt.BackwardOutConstraint

DiffOpt.BackwardOutObjective

DiffOpt.ForwardInConstraint

DiffOpt.ForwardInObjective

DiffOpt.ForwardOutVariablePrimal

DiffOpt.IndexMappedFunction

DiffOpt.MOItoJuMP

DiffOpt.MatrixScalarQuadraticFunction

DiffOpt.MatrixVectorAffineFunction

DiffOpt.ProductOfSets

DiffOpt.ProgramClass

DiffOpt.ProgramClassCode

DiffOpt.ProgramClassUsed

DiffOpt.VectorScalarAffineFunction

DiffOpt.Dπ

DiffOpt._backward_conic

DiffOpt._backward_quad

DiffOpt._forward_conic

DiffOpt._forward_quad

DiffOpt.backward

DiffOpt.create_LHS_matrix

DiffOpt.diff_optimizer

DiffOpt.diff_optimizer

DiffOpt.forward

DiffOpt.get_problem_data

DiffOpt.map_rows

DiffOpt.quad_sym_half

DiffOpt.standard_form

DiffOpt.π

DiffOpt.AbstractLazyScalarFunction — Type

abstract type AbstractLazyScalarFunction <: MOI.AbstractScalarFunction end

Subtype of MOI.AbstractScalarFunction that is not a standard MOI scalar function but can be converted to one using standard_form.

The function can also be inspected lazily using JuMP.coefficient or quad_sym_half.

DiffOpt.BackwardInVariablePrimal — Type

BackwardInVariablePrimal <: MOI.AbstractVariableAttribute

A MOI.AbstractVariableAttribute to set input data to backward differentiation, that is, problem solution.

For instance, to set the tangent of the variable of index vi, do the following:

MOI.set(model, DiffOpt.BackwardInVariablePrimal(), x)

DiffOpt.BackwardOutConstraint — Type

BackwardOutConstraint

An MOI.AbstractConstraintAttribute to get output data to backward differentiation, that is, problem input data.

For instance, if the following returns x + 2y + 5, it means that the tangent has coordinate 1 for the coefficient of x, coordinate 2 for the coefficient of y and 5 for the function constant. If the constraint is of the form func == constant or func <= constant, the tangent for the constant on the right-hand side is -5.

MOI.get(model, DiffOpt.BackwardOutConstraint(), ci)

DiffOpt.BackwardOutObjective — Type

BackwardOutObjective <: MOI.AbstractModelAttribute

A MOI.AbstractModelAttribute to get output data to backward differentiation, that is, problem input data.

For instance, to get the tangent of the objective function corresponding to the tangent given to BackwardInVariablePrimal, do the following:

func = MOI.get(model, DiffOpt.BackwardOutObjective)

Then, to get the sensitivity of the linear term with variable x, do

JuMP.coefficient(func, x)

To get the sensitivity with respect to the quadratic term with variables x and y, do either

JuMP.coefficient(func, x, y)

or

DiffOpt.quad_sym_half(func, x, y)

Warning

These two lines are not equivalent in case x == y, see quad_sym_half for the details on the difference between these two functions.

DiffOpt.ForwardInConstraint — Type

ForwardInConstraint <: MOI.AbstractConstraintAttribute

A MOI.AbstractConstraintAttribute to set input data to forward differentiation, that is, problem input data.

For instance, if the scalar constraint of index ci contains θ * (x + 2y) <= 5θ, for the purpose of computing the derivative with respect to θ, the following should be set:

MOI.set(model, DiffOpt.ForwardInConstraint(), ci, 1.0 * x + 2.0 * y - 5.0)

Note that we use -5 as the ForwardInConstraint sets the tangent of the ConstraintFunction so we consider the expression θ * (x + 2y - 5).

DiffOpt.ForwardInObjective — Type

ForwardInObjective <: MOI.AbstractModelAttribute

A MOI.AbstractModelAttribute to set input data to forward differentiation, that is, problem input data. The possible values are any MOI.AbstractScalarFunction. A MOI.ScalarQuadraticFunction can only be used in linearly constrained quadratic models.

For instance, if the objective contains θ * (x + 2y), for the purpose of computinig the derivative with respect to θ, the following should be set:

MOI.set(model, DiffOpt.ForwardInObjective(), 1.0 * x + 2.0 * y)

where x and y are the relevant MOI.VariableIndex.

DiffOpt.ForwardOutVariablePrimal — Type

ForwardOutVariablePrimal <: MOI.AbstractVariableAttribute

A MOI.AbstractVariableAttribute to get output data from forward differentiation, that is, problem solution.

For instance, to get the tangent of the variable of index vi corresponding to the tangents given to ForwardInObjective and ForwardInConstraint, do the following:

MOI.get(model, DiffOpt.ForwardOutVariablePrimal(), vi)

DiffOpt.IndexMappedFunction — Type

IndexMappedFunction{F<:MOI.AbstractFunction} <: AbstractLazyScalarFunction

Lazily represents the function MOI.Utilities.map_indices(index_map, DiffOpt.standard_form(func)).

DiffOpt.MOItoJuMP — Type

MOItoJuMP{F<:MOI.AbstractScalarFunction} <: JuMP.AbstractJuMPScalar

Lazily represents the function JuMP.jump_function(model, DiffOpt.standard_form(func)).

DiffOpt.MatrixScalarQuadraticFunction — Type

struct MatrixScalarQuadraticFunction{T, VT, MT} <: MOI.AbstractScalarFunction
    affine::VectorScalarAffineFunction{T,VT}
    terms::MT
end

Represents the function x' * terms * x / 2 + affine as an MOI.AbstractScalarFunction where x[i] = MOI.VariableIndex(i). Use standard_form to convert it to a MOI.ScalarQuadraticFunction{T}.

DiffOpt.MatrixVectorAffineFunction — Type

MatrixVectorAffineFunction{T, VT} <: MOI.AbstractVectorFunction

Represents the function terms * x + constant as an MOI.AbstractVectorFunction where x[i] = MOI.VariableIndex(i). Use standard_form to convert it to a MOI.VectorAffineFunction{T}.

DiffOpt.ProductOfSets — Type

ProductOfSets{T} <: MOI.Utilities.OrderedProductOfSets{T}

The MOI.Utilities.@product_of_sets macro requires to know the list of sets at compile time. In DiffOpt however, the list depends on what the user is going to use as set as DiffOpt supports any set as long as it implements the required function of MathOptSetDistances. For this type, the list of sets can be given a run-time.

DiffOpt.ProgramClass — Type

ProgramClass <: MOI.AbstractOptimizerAttribute

Determines which program class to used from ProgramClassCode. The default is AUTOMATIC.

One important advantage of setting the class explicitly is that it will allow necessary bridges to be used. If the class is AUTOMATIC then DiffOpt.Optimizer will report that it supports both objective and constraints of the QP and CP classes. For instance, it will reports that is supports both quadratic objective and conic constraints. However, at the differentiation stage, we won't be able to differentiate since QP does not support conic constraints and CP does not support quadratic objective. On the other hand, if the ProgramClass is set to CONIC then DiffOpt.Optimizer will report that it does not support quadratic objective hence it will be bridged to second-order cone constraints and we will be able to use CP to differentiate.

DiffOpt.ProgramClassCode — Type

@enum ProgramClassCode QUADRATIC CONIC AUTOMATIC

Program class used by DiffOpt. DiffOpt implements differentiation of two different program class:

Quadratic Program (QP): quadratic objective and linear constraints and

Conic Program (CP): linear objective and conic constraints.

AUTOMATIC which means that the class will be automatically selected given the problem data: if any constraint is conic, CP is used and QP is used otherwise. See ProgramClass.

DiffOpt.ProgramClassUsed — Type

ProgramClassUsed <: MOI.AbstractOptimizerAttribute

Program class actually used, same as ProgramClass except that it does not return AUTOMATIC but the class automatically chosen instead. This attribute is read-only, it cannot be set, set ProgramClass instead.

DiffOpt.VectorScalarAffineFunction — Type

VectorScalarAffineFunction{T, VT} <: MOI.AbstractScalarFunction

Represents the function x ⋅ terms + constant as an MOI.AbstractScalarFunction where x[i] = MOI.VariableIndex(i). Use standard_form to convert it to a MOI.ScalarAffineFunction{T}.

DiffOpt.Dπ — Method

Dπ(v::Vector{Float64}, model, cones::ProductOfSets, index_map::MOIU.IndexMap)

Given a model, its cones and the index_map from the indices of model to the indices of cones, find the gradient of the projection of the vectors v of length equal to the number of rows in the conic form onto the cartesian product of the cones corresponding to these rows. For more info, refer to https://github.com/matbesancon/MathOptSetDistances.jl

DiffOpt._backward_conic — Method

_backward_conic(model::Optimizer, dx::Vector{Float64}, dy::Vector{Float64}, ds::Vector{Float64})

Method to compute the product of the transpose of the derivative (Jacobian) at the conic program parameters A, b, c to the perturbations dx, dy, ds. This is similar to backward.

For theoretical background, refer Section 3 of Differentiating Through a Cone Program, https://arxiv.org/abs/1904.09043

DiffOpt._backward_quad — Method

_backward_quad(model::Optimizer)

Method to differentiate optimal solution z and return product of jacobian matrices (dz / dQ, dz / dq, etc) with the backward pass vector dl / dz

The method computes the product of

jacobian of problem solution z* with respect to problem parameters set with the BackwardInVariablePrimal

a backward pass vector dl / dz, where l can be a loss function

Note that this method does not returns the actual jacobians.

For more info refer eqn(7) and eqn(8) of https://arxiv.org/pdf/1703.00443.pdf

DiffOpt._forward_conic — Method

_forward_conic(model::Optimizer)

Method to compute the product of the derivative (Jacobian) at the conic program parameters A, b, c to the perturbations dA, db, dc. This is similar to forward.

For theoretical background, refer Section 3 of Differentiating Through a Cone Program, https://arxiv.org/abs/1904.09043

DiffOpt._forward_quad — Method

_forward_quad(model::Optimizer)

DiffOpt.backward — Method

backward(model::Optimizer)

Wrapper method for the backward pass. This method will consider as input a currently solved problem and differentials with respect to the solution set with the BackwardInVariablePrimal attribute. The output problem data differentials can be queried with the attributes BackwardOutObjective and BackwardOutConstraint.

DiffOpt.create_LHS_matrix — Function

create_LHS_matrix(z, λ, Q, G, h, A=nothing)

Inverse matrix specified on RHS of eqn(7) in https://arxiv.org/pdf/1703.00443.pdf

Helper method while calling _backward_quad

DiffOpt.diff_optimizer — Method

diff_optimizer(optimizer_constructor)::Optimizer

Creates a DiffOpt.Optimizer, which is an MOI layer with an internal optimizer and other utility methods. Results (primal, dual and slack values) are obtained by querying the internal optimizer instantiated using the optimizer_constructor. These values are required for find jacobians with respect to problem data.

One define a differentiable model by using any solver of choice. Example:

julia> using DiffOpt, GLPK

julia> model = diff_optimizer(GLPK.Optimizer)
julia> model.add_variable(x)
julia> model.add_constraint(...)

julia> _backward_quad(model)  # for convex quadratic models

julia> _backward_quad(model)  # for convex conic models

DiffOpt.forward — Method

forward(model::Optimizer)

Wrapper method for the forward pass. This method will consider as input a currently solved problem and differentials with respect to problem data set with the ForwardInObjective and ForwardInConstraint attributes. The output solution differentials can be queried with the attribute ForwardOutVariablePrimal.

DiffOpt.get_problem_data — Method

get_problem_data(model::MOI.AbstractOptimizer)

Return problem parameters as matrices along with other program info such as number of constraints, variables, etc

DiffOpt.map_rows — Method

map_rows(f::Function, model, cones::ProductOfSets, index_map::MOIU.IndexMap, map_mode::Union{Nested{T}, Flattened{T}})

Given a model, its cones, the index_map from the indices of model to the indices of cones and map_mode of type Nested (resp. Flattened), return a Vector{T} of length equal to the number of cones (resp. rows) in the conic form where the value for the index (resp. rows) corresponding to each cone is equal to f(ci, r) where ci is the corresponding constraint index in model and r is a UnitRange of the corresponding rows in the conic form.

DiffOpt.quad_sym_half — Function

quad_sym_half(func, vi1::MOI.VariableIndex, vi2::MOI.VariableIndex)

Return Q[i,j] = Q[j,i] where the quadratic terms of func is represented by x' Q x / 2 for a symmetric matrix Q where x[i] = vi1 and x[j] = vi2. Note that while this is equal to JuMP.coefficient(func, vi1, vi2) if vi1 != vi2, in the case vi1 == vi2, it is rather equal to 2JuMP.coefficient(func, vi1, vi2).

DiffOpt.standard_form — Function

standard_form(func::AbstractLazyScalarFunction)

Converts func to a standard MOI scalar function.

standard_form(func::MOItoJuMP)

Converts func to a standard JuMP scalar function.

DiffOpt.π — Method

π(v::Vector{Float64}, model::MOI.ModelLike, cones::ProductOfSets, index_map::MOIU.IndexMap)

Given a model, its cones and the index_map from the indices of model to the indices of cones, find the projection of the vectors v of length equal to the number of rows in the conic form onto the cartesian product of the cones corresponding to these rows. For more info, refer to https://github.com/matbesancon/MathOptSetDistances.jl